+++ scripts/utils/__init__.py
@@ -1,3 +1,4 @@
-from .system import *
-from .dates import *
+from .system import check_folder_or_create, cls
+from .dates import get_now_as_str
 from .requests import Requester
+from .numbers import pad