+++ scripts/utils/numbers.py
@@ -1,0 +1,2 @@
+def pad(number: int, padding: int = 3) -> str:
+    return str(number).zfill(padding)